feat(plug-in): insert bundle before preprocessed files
This commit tries to figure out the correct position of the bundle file
based on config.files and config.preprocessors.

By doing so, it makes sure that external (non-commonJS) files can be
inserted before the bundle file and get correctly discovered.

It compares defined files against browserified files defined in
config.preprocessors and inserts the bundle in the first matching file
found. If none is found, it defaults to the old behavior. That is, it
inserts the bundle in the file list at position zero.

Closes #1