fix(travis): make node version a string
This is a common mistake with Travis files. It used to not cause problems but apparently they now changed something. YAML parses `0.10` to `0.1`. Either Travis or nvm used to assume this meant `0.10` but apparently now it breaks. Strings are always the right convention anyhow since version numbers aren't actually numbers.