fix(plug-in): revert conditional rebuild on change
This broke change detection for external resources included via brfs.

Instead of relying on the logic we must rely on debouncing the bundle
operation to suppress the double bundle.

The delay before an actual bundling operation is triggered may now be
configured via a browserify bundleDelay property in the karma
configuration.

Related to #69