fix(bro): remove bundleDelay from watchify config
By propagating the delay to watchify, the rebundling will most likely happen
twice. This is due to the fact that karma-browserify waits for
timing(fs event + watchify update) which has a high likelyhood to exceed
timing(fs event).

Putting the watchify delay back to default seems to resolve the rebundling
issues on my machine.

Closes #116